Add files via upload
This internship was contain around two main tasks to accomplished. the Task_01 was based on a synthesised transaction dataset containing 3 months’ worth of transactions for 100 hypothetical customers. It contains purchases, recurring transactions, and salary transactions. my job was to loading and cleaning the dataset and gathering some interesting overall insights about the data. on Task_02 the main duty was based on visualizing any correlation by using a scatter plot and also building a simple regression model to predict the annual salary for each customer.